pyproject.toml
[project]
name = "autoresearch-rivals"
version = "0.1.0"
description = "Adversarial dual-agent research system — fork of autoresearch"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.0",
    "tiktoken>=0.11.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
    "pymupdf>=1.23.0",
    "sentence-transformers>=2.2.0",
    "rich>=13.0.0",
    "click>=8.1.0",
]
